Add: Complete task 4 of Python Classes
A change was also made to 3-square.py. Prevoiusly, the area method had
an instance varaible by the name 'area'. This would cause a conflict
when using the name in a call since the function has a similar name.
This would result in an error since there would be ambiguity